requirements.txt
langchain
chromadb
sentence_transformers
python-dotenv
langchain_groq
ragas
bert_score
rouge
nltk
pydantic
langchain_core
langchain_community
llama_index
llama-index-agent-introspective
llama-index-llms-groq
langchain_experimental
cohere
langgraph
llama-index-embeddings-huggingface
llama-index-embeddings-instructor
llama-index-tools-tavily-research
tavily-python
llama-hub
